继续推进Spec(Specific)体系，完成了最基本的提交Task和根据TaskId删除任务

+++ Cherry-Core/src/main/java/cn/cherry/core/engine/base/executor/ScheduleExecutor.java
@@ -17,10 +17,10 @@ import java.util.concurrent.BlockingQueue;
  */
 public class ScheduleExecutor implements Runnable {
 
-    private final int interval;
+    private final long interval;
     private final BlockingQueue<Integer> messageChannel;
 
-    public ScheduleExecutor(int interval, BlockingQueue<Integer> messageChannel) {
+    public ScheduleExecutor(long interval, BlockingQueue<Integer> messageChannel) {
         this.interval = interval;
         this.messageChannel = messageChannel;
     }
